Move jemalloc to metrics to a sepearte file, and load from app to get proper logs
